requirements.txt
flask
Pillow
exif
flask-sqlalchemy
flask-login
werkzeug
python-dotenv 